Motion profile testing WIP
the profile runs, but it runs too fast, and as a result the position is far behind what it should be, and the profile runs for far longer to try and match that.

at the moment it runs, but wierd, need to figure out the specifics on how to run this motion profile stuff.