Added Feathering to Brake
need to tune the brake on the feathering depending on slope

need to run motor at a speed, then feather at multiple rates, and check if its a linear decrease from feather percent to rate of decrease.

also code is all untested, so need to make sure that the brake feathering code actually works.