Worked on the new braking algorithm
turns out a lot of the issues i got were divide by 0, so thats suboptimal, but that got fixed

ultimately, i ended up making the braking mode that zach and maziar wanted, with a brake count and a duration for each count.

if you hit the left bumper, you can also toggle manual mode, which lets you use the B button to deactivate the brake manually.